- Removed unnecessary code from examples that use shadow maps.
- Unittests cleanup.
- Replaced the XR_ADVANCE_PTR() macro with a constexpr function, xr::meshutil::AdvancePointer().
- xr::Quaternion and xr::Matrix API improvements.
- xr::Quaternion fixes and improvements.
- Header cleanup in xr3core math.
- Fix: xr::detail::QueueIterator<>::operator++(int) returns copy, not dangling reference.
- XonEntity modernisation / improvements.
- Added <stdexcept> #includes for GCC10 builds.
- xr::Counted(Core) fixes / improvements;

+++ unittests/src/SpinlockTests.cpp
@@ -22,13 +22,13 @@ struct Data
 {
   Spinlock spinner;
 
-  int accessCounter = 0;
+  size_t accessCounter = 0;
   std::vector<int> data;
 
   std::chrono::time_point<std::chrono::system_clock> tStart;
 
   Data(size_t n)
-    : data(n),
+  : data(n),
     tStart(std::chrono::system_clock::now())
   {}
